requirements.txt
# Прямые зависимости runtime бота (production / ручной запуск python bot.py).
# Транзитивные версии можно зафиксировать в requirements-lock.txt.
# Для pytest см. requirements-test.txt.

matrix-nio>=0.25.0          # Matrix API — отправка сообщений в чат
python-redmine>=2.5.0       # Redmine API — получение задач и журналов
python-dotenv>=1.0.0        # Загрузка переменных из .env файла
APScheduler>=3.10.0         # Планировщик периодических задач

# Админка и Postgres (конфиг в БД; state пока в JSON)
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.4
python-multipart>=0.0.17
alembic>=1.14.0
psycopg[binary]>=3.2.0      # sync Alembic миграции
argon2-cffi>=23.1.0
cryptography>=43.0.0
httpx>=0.27.0               # HTTP-клиент для heartbeat-запросов к админке
cachetools>=5.5.0           # TTLCache для кэша админки
python-json-logger>=3.2.0   # Structured JSON logging
